requirements.txt
patool
tinyscript>=1.23.16
